backend/refresh_token/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.3.2
functions-framework==3.3.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
python-dotenv==1.0.0
requests==2.30.0
urllib3==2.0.2
watchdog==3.0.0
Werkzeug==2.3.4
zipp==3.15.0
